infra/cdk/requirements.txt
as==0.1
attrs==20.3.0
autopep8==1.5.7
aws-cdk.assets
aws-cdk.aws-apigateway==1.106.1
aws-cdk.aws-apigatewayv2==1.106.1
aws-cdk.aws-applicationautoscaling
aws-cdk.aws-autoscaling==1.106.1
aws-cdk.aws-autoscaling-common
aws-cdk.aws-autoscaling-hooktargets==1.106.1
aws-cdk.aws-certificatemanager
aws-cdk.aws-cloudformation
aws-cdk.aws-cloudfront
aws-cdk.aws-cloudfront-origins
aws-cdk.aws-cloudwatch==1.106.1
aws-cdk.aws-codebuild==1.106.1
aws-cdk.aws-codecommit==1.106.1
aws-cdk.aws-codedeploy==1.106.1
aws-cdk.aws-codeguruprofiler
aws-cdk.aws-codepipeline==1.106.1
aws-cdk.aws-codepipeline-actions==1.106.1
aws-cdk.aws-cognito==1.106.1
aws-cdk.aws-config==1.106.1
aws-cdk.aws-dynamodb==1.106.1
aws-cdk.aws-ec2
aws-cdk.aws-ecr
aws-cdk.aws-ecr-assets
aws-cdk.aws-ecs==1.106.1
aws-cdk.aws-efs
aws-cdk.aws-elasticloadbalancing==1.106.1
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-events
aws-cdk.aws-events-targets==1.106.1
aws-cdk.aws-globalaccelerator==1.106.1
aws-cdk.aws-iam
aws-cdk.aws-kinesis==1.106.1
aws-cdk.aws-kinesisfirehose==1.106.1
aws-cdk.aws-kms
aws-cdk.aws-lambda
aws-cdk.aws-logs
aws-cdk.aws-route53
aws-cdk.aws-route53-targets==1.106.1
aws-cdk.aws-s3
aws-cdk.aws-s3-assets
aws-cdk.aws-sam==1.106.1
aws-cdk.aws-secretsmanager==1.106.1
aws-cdk.aws-servicediscovery==1.106.1
aws-cdk.aws-signer
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions==1.106.1
aws-cdk.aws-sqs
aws-cdk.aws-ssm
aws-cdk.aws-stepfunctions==1.106.1
aws-cdk.cloud-assembly-schema
aws-cdk.core
aws-cdk.custom-resources
aws-cdk.cx-api
aws-cdk.region-info
boto3==1.17.83
botocore==1.20.83
cattrs==1.5.0
constructs==3.3.75
dynamod==0.8.4
gradescent==0.9.4
jmespath==0.10.0
jsii==1.29.0
publication==0.0.3
pycodestyle==2.7.0
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
